requirements.txt
requests
cos-python-sdk-v5
grpcio
grpcio-tools
cachetools
urllib3
tcvdb-text